Changed to integer

+++ backend/db/migrations/20230807231102-create-review.js
@@ -28,7 +28,7 @@ module.exports = {
         allowNull: false
       },
       stars: {
-        type: Sequelize.DECIMAL(2,1),
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       createdAt: {
